fourth commit

+++ routes/web.php
@@ -12,152 +12,190 @@ Route::get('/', function () {
 del sistema según el siste map <---*/
 
 /*Ruta principal del sistemas*/
-Route::get('/', [RutasController::class, 'inicio' ]); 
+Route::get('/', [RutasController::class, 'inicio' ])->name('inicio'); 
+
+
+/*--->*/
+Route::get('home', [RutasController::class, 'home' ])->name('home'); 
+
 /*--->*/
 
-/*Ruta para la vista administrativa*/
+/*En esta ruta obtenemos el listado de todas las fincas*/
 
-Route::get('/home/sisga-admin', [App\Http\Controllers\HomeController::class, 'admin'])->name('admin'); 
-/*--->*/
-
-/*Ruta para la vista finca*/
-
-Route::get('/home/sisga-admin/fincas', [App\Http\Controllers\HomeController::class, 'fincas'])->name('fincas');
+Route::get('/home/fincas', [App\Http\Controllers\HomeController::class, 'fincas'])->name('fincas');
 
 /*Con ésta ruta pasamos los argumentos al controlador registro y guardar en la tabla finca */
-Route::post('/home/sisga-admin/fincas', [App\Http\Controllers\HomeController::class, 'crear'])->name('fincas.crear'); 
+Route::post('/home/fincas', [App\Http\Controllers\HomeController::class, 'crear'])->name('fincas.crear'); 
 /*Con ésta ruta pasamos los argumentos al controlador para editar el registro y guardar en la tabla finca */
-Route::get('/home/sisga-admin/fincas/{id}', [App\Http\Controllers\HomeController::class, 'editar'])->name('fincas.editar'); 
+Route::get('/home/fincas/{id}', [App\Http\Controllers\HomeController::class, 'editar'])->name('fincas.editar'); 
 
-Route::put('/home/sisga-admin/fincas/{id}', [App\Http\Controllers\HomeController::class, 'update'])->name('fincas.update');
+Route::put('/home/fincas/{id}', [App\Http\Controllers\HomeController::class, 'update'])->name('fincas.update');
 
-Route::delete('/home/sisga-admin/fincas/eliminar/{id}', [App\Http\Controllers\HomeController::class, 'eliminar'])->name('fincas.eliminar'); 
+Route::delete('/home/fincas/eliminar/{id}', [App\Http\Controllers\HomeController::class, 'eliminar'])->name('fincas.eliminar'); 
 
 /*-->Fin Rutas Finca*/
 
 
-//-> Inicio Ruta de Variable de Control Especie
+/*Ruta para la vista administrativa*/
 
+Route::get('/home/sisga-admin/finca/{id}', [App\Http\Controllers\HomeController::class, 'admin'])->name('admin'); 
+/*--->*/
+
+/***********************************************
+* Inicio Ruta de Variable de Control Especie
+************************************************/
+
+Route::get('/home/sisga-admin/finca/{id_finca}/especie', [App\Http\Controllers\HomeController::class, 'especie'])->name('especie');
+/*
 Route::get('/home/sisga-admin/especie', [App\Http\Controllers\HomeController::class, 'especie'])->name('especie');
+*/
+Route::post('/home/sisga-admin/finca/{id_finca}/especie', [App\Http\Controllers\HomeController::class, 'crear_especie'])->name('especie.crear');
 
-//->> EndEspecie 
+Route::get('/home/sisga-admin/finca/{id_finca}/especie/{id}', [App\Http\Controllers\HomeController::class, 'editar_especie'])->name('especie.editar'); 
 
+Route::put('/home/sisga-admin/finca/{id_finca}/especie/{id}', [App\Http\Controllers\HomeController::class, 'update_especie'])->name('especie.update');
+
+Route::delete('/home/sisga-admin/finca/{id_finca}/especie/{id}', [App\Http\Controllers\HomeController::class, 'eliminar_especie'])->name('especie.eliminar'); 
+
+/***********************************************
+* Fin Ruta de Variable de Control Especie
+************************************************/
+
+/***********************************************
+* Inicio Ruta de Variable de Control raza
+************************************************/
+
+Route::get('/home/sisga-admin/finca/{id_finca}/raza', [App\Http\Controllers\HomeController::class, 'raza'])->name('raza');
+/*
+Route::get('/home/sisga-admin/raza', [App\Http\Controllers\HomeController::class, 'raza'])->name('raza');
+*/
+
+Route::post('/home/sisga-admin/finca/{id_finca}/raza', [App\Http\Controllers\HomeController::class, 'crear_raza'])->name('raza.crear'); 
+
+Route::get('/home/sisga-admin/finca/{id_finca}/raza/{id}', [App\Http\Controllers\HomeController::class, 'editar_raza'])->name('raza.editar'); 
+
+Route::put('/home/sisga-admin/finca/{id_finca}/raza/{id}', [App\Http\Controllers\HomeController::class, 'update_raza'])->name('raza.update');
+
+Route::delete('/home/sisga-admin/finca/{id_finca}/raza/{id}', [App\Http\Controllers\HomeController::class, 'eliminar_raza'])->name('raza.eliminar'); 
+
+/***********************************************
+* End Ruta de Variable de Control raza
+************************************************/
+
+
+/***********************************************
+* Inicio Ruta de Variable de Control Tipologia
+************************************************/
 /*Ruta para la vista tipologia*/
-
+Route::get('/home/sisga-admin/finca/{id_finca}/tipologias', [App\Http\Controllers\HomeController::class, 'tipologia'])->name('tipologia');
+/*
 Route::get('/home/sisga-admin/tipologias', [App\Http\Controllers\HomeController::class, 'tipologia'])->name('tipologia');
+*/
 
-Route::post('/home/sisga-admin/tipologias', [App\Http\Controllers\HomeController::class, 'crear_tipo'])->name('tipologia.crear'); 
+Route::post('/home/sisga-admin/finca/{id_finca}/tipologias', [App\Http\Controllers\HomeController::class, 'crear_tipo'])->name('tipologia.crear'); 
 
-Route::get('/home/sisga-admin/tipologias/{id}', [App\Http\Controllers\HomeController::class, 'editar_tipo'])->name('tipologia.editar'); 
+Route::get('/home/sisga-admin/finca/{id_finca}/tipologias/{id}', [App\Http\Controllers\HomeController::class, 'editar_tipo'])->name('tipologia.editar'); 
 
-Route::put('/home/sisga-admin/tipologias/{id}', [App\Http\Controllers\HomeController::class, 'update_tipo'])->name('tipologia.update');
+Route::put('/home/sisga-admin/finca/{id_finca}/tipologias/{id}', [App\Http\Controllers\HomeController::class, 'update_tipo'])->name('tipologia.update');
 
-Route::delete('/home/sisga-admin/tipologias/eliminar/{id}', [App\Http\Controllers\HomeController::class, 'eliminar_tipo'])->name('tipologia.eliminar'); 
+Route::delete('/home/sisga-admin/finca/{id_finca}/tipologias/{id}', [App\Http\Controllers\HomeController::class, 'eliminar_tipo'])->name('tipologia.eliminar'); 
 
-/*End Route Tipologia */
+/***********************************************
+* End Ruta de Variable de Control Tipologia
+************************************************/
+
+/***********************************************
+* Inicio de Ruta para la vista Condiciones Corporales
+************************************************/
 
 
-/*---> Ruta para la vista Condiciones Corporales*/
-
+Route::get('/home/sisga-admin/finca/{id_finca}/condicion-corporal', [App\Http\Controllers\HomeController::class, 'condicion_corporal'])->name('condicion_corporal');
+/*
 Route::get('/home/sisga-admin/condicion-corporal', [App\Http\Controllers\HomeController::class, 'condicion_corporal'])->name('condicion_corporal');
+*/
+Route::post('/home/sisga-admin/finca/{id_finca}/condicion-corporal', [App\Http\Controllers\HomeController::class, 'crear_condicioncorporal'])->name('condicion_corporal.crear'); 
 
-Route::post('/home/sisga-admin/condicion-corporal', [App\Http\Controllers\HomeController::class, 'crear_condicioncorporal'])->name('condicion_corporal.crear'); 
+Route::get('/home/sisga-admin/finca/{id_finca}/condicion-corporal/{id}', [App\Http\Controllers\HomeController::class, 'editar_condicion'])->name('condicion_corporal.editar'); 
 
-Route::get('/home/sisga-admin/condicion-corporal/{id}', [App\Http\Controllers\HomeController::class, 'editar_condicion'])->name('condicion_corporal.editar'); 
+Route::put('/home/sisga-admin/finca/{id_finca}/condicion-corporal/{id}', [App\Http\Controllers\HomeController::class, 'update_condicion'])->name('condicion_corporal.update');
 
-Route::put('/home/sisga-admin/condicion-corporal/{id}', [App\Http\Controllers\HomeController::class, 'update_condicion'])->name('condicion_corporal.update');
+Route::delete('/home/sisga-admin/finca/{id_finca}/condicion-corporal/{id}', [App\Http\Controllers\HomeController::class, 'eliminar_condicion'])->name('condicion_corporal.eliminar'); 
 
-Route::delete('/home/sisga-admin/condicion-corporal/eliminar/{id}', [App\Http\Controllers\HomeController::class, 'eliminar_condicion'])->name('condicion_corporal.eliminar'); 
+/***********************************************
+* End de Ruta para la vista Condiciones Corporales
+************************************************/
 
-/*End Ruta para Condiciones Corporales*/
+/***********************************************
+* *Ruta para el diagnostico de palpaciones
+************************************************/
 
+Route::get('/home/sisga-admin/finca/{id_finca}/diagnostico-palpaciones', [App\Http\Controllers\HomeController::class, 'diagnostico_palpaciones'])->name('diagnostico_palpaciones');
+/*
+Route::get('/home/sisga-admin/diagnostico-palpaciones', [App\Http\Controllers\HomeController::class, 'diagnostico_palpaciones'])->name('diagnostico_palpaciones'); */
 
+Route::post('/home/sisga-admin/finca/{id_finca}/diagnostico-palpaciones', [App\Http\Controllers\HomeController::class, 'crear_diagnosticopalpaciones'])->name('diagnostico_palpaciones.crear'); 
 
-/*Ruta para el diagnostico de palpaciones*/
+Route::get('/home/sisga-admin/finca/{id_finca}/diagnostico-palpaciones/{id}', [App\Http\Controllers\HomeController::class, 'editar_diagnostico_palpaciones'])->name('diagnostico_palpaciones.editar'); 
 
-Route::get('/home/sisga-admin/diagnostico-palpaciones', [App\Http\Controllers\HomeController::class, 'diagnostico_palpaciones'])->name('diagnostico_palpaciones');
+Route::put('/home/sisga-admin/finca/{id_finca}/diagnostico-palpaciones/{id}', [App\Http\Controllers\HomeController::class, 'update_diagnostico_palpaciones'])->name('diagnostico_palpaciones.update');
 
-Route::post('/home/sisga-admin/diagnostico-palpaciones', [App\Http\Controllers\HomeController::class, 'crear_diagnosticopalpaciones'])->name('diagnostico_palpaciones.crear'); 
-
-Route::get('/home/sisga-admin/diagnostico-palpaciones/{id}', [App\Http\Controllers\HomeController::class, 'editar_diagnostico_palpaciones'])->name('diagnostico_palpaciones.editar'); 
-
-Route::put('/home/sisga-admin/diagnostico-palpaciones/{id}', [App\Http\Controllers\HomeController::class, 'update_diagnostico_palpaciones'])->name('diagnostico_palpaciones.update');
-
-Route::delete('/home/sisga-admin/diagnostico-palpaciones/eliminar/{id}', [App\Http\Controllers\HomeController::class, 'eliminar_diagnostico_palpaciones'])->name('diagnostico_palpaciones.eliminar'); 
+Route::delete('/home/sisga-admin/finca/{id_finca}/diagnostico-palpaciones/{id}', [App\Http\Controllers\HomeController::class, 'eliminar_diagnostico_palpaciones'])->name('diagnostico_palpaciones.eliminar'); 
 /*-->> End Ruta palpaciones*/
 
 
 
 /*Ruta para motivos de entrada y salidas de series*/
+
+Route::get('/home/sisga-admin/finca/{id_finca}/motivos-entrada-salida', [App\Http\Controllers\HomeController::class, 'motivo_entrada_salida'])->name('motivo_entrada_salida');
+/*
 Route::get('/home/sisga-admin/motivos-entrada-salida', [App\Http\Controllers\HomeController::class, 'motivo_entrada_salida'])->name('motivo_entrada_salida');
+*/
+Route::post('/home/sisga-admin/finca/{id_finca}/motivos-entrada-salida', [App\Http\Controllers\HomeController::class, 'crear_motivo_entrada_salida'])->name('motivo_entrada_salida.crear'); 
 
-Route::post('/home/sisga-admin/motivos-entrada-salida', [App\Http\Controllers\HomeController::class, 'crear_motivo_entrada_salida'])->name('motivo_entrada_salida.crear'); 
+Route::get('/home/sisga-admin/finca/{id_finca}/motivos-entrada-salida/{id}', [App\Http\Controllers\HomeController::class, 'editar_motivo_entrada_salida'])->name('motivo_entrada_salida.editar'); 
 
-Route::get('/home/sisga-admin/motivos-entrada-salida/{id}', [App\Http\Controllers\HomeController::class, 'editar_motivo_entrada_salida'])->name('motivo_entrada_salida.editar'); 
+Route::put('/home/sisga-admin/finca/{id_finca}/motivos-entrada-salida/{id}', [App\Http\Controllers\HomeController::class, 'update_motivo_entrada_salida'])->name('motivo_entrada_salida.update');
 
-Route::put('/home/sisga-admin/motivos-entrada-salida/{id}', [App\Http\Controllers\HomeController::class, 'update_motivo_entrada_salida'])->name('motivo_entrada_salida.update');
-
-Route::delete('/home/sisga-admin/motivos-entrada-salida/eliminar/{id}', [App\Http\Controllers\HomeController::class, 'eliminar_motivo_entrada_salida'])->name('motivo_entrada_salida.eliminar'); 
+Route::delete('/home/sisga-admin/finca/{id_finca}/motivos-entrada-salida/{id}', [App\Http\Controllers\HomeController::class, 'eliminar_motivo_entrada_salida'])->name('motivo_entrada_salida.eliminar'); 
 /*-->> End Ruta motivos*/
 
 
 
-/*Ruta para el registro de Razas*/
 
-Route::get('/home/sisga-admin/raza', [App\Http\Controllers\HomeController::class, 'raza'])->name('raza');
-
-Route::post('/home/sisga-admin/raza', [App\Http\Controllers\HomeController::class, 'crear_raza'])->name('raza.crear'); 
-
-Route::get('/home/sisga-admin/raza/{id}', [App\Http\Controllers\HomeController::class, 'editar_raza'])->name('raza.editar'); 
-
-Route::put('/home/sisga-admin/raza/{id}', [App\Http\Controllers\HomeController::class, 'update_raza'])->name('raza.update');
-
-Route::delete('/home/sisga-admin/raza/eliminar/{id}', [App\Http\Controllers\HomeController::class, 'eliminar_raza'])->name('raza.eliminar'); 
-
-/*-->> End Ruta raza*/
 
 /*Ruta para el registro de patologías*/
-
-Route::get('/home/sisga-admin/patologia', [App\Http\Controllers\HomeController::class, 'patologia'])->name('patologia');
-
-Route::post('/home/sisga-admin/patologia', [App\Http\Controllers\HomeController::class, 'crear_patologia'])->name('patologia.crear');
-
-Route::get('/home/sisga-admin/patologia/{id}', [App\Http\Controllers\HomeController::class, 'editar_patologia'])->name('patologia.editar'); 
-
-Route::put('/home/sisga-admin/patologia/{id}', [App\Http\Controllers\HomeController::class, 'update_patologia'])->name('patologia.update');
-
-Route::delete('/home/sisga-admin/patologia/eliminar/{id}', [App\Http\Controllers\HomeController::class, 'eliminar_patologia'])->name('patologia.eliminar'); 
-
-
-
+Route::get('/home/sisga-admin/finca/{id_finca}/patologia', [App\Http\Controllers\HomeController::class, 'patologia'])->name('patologia');
 /*
-Route::post('/home/sisga-admin/raza', [App\Http\Controllers\HomeController::class, 'crear_raza'])->name('raza.crear'); 
+Route::get('/home/sisga-admin/patologia', [App\Http\Controllers\HomeController::class, 'patologia'])->name('patologia');
+*/
+Route::post('/home/sisga-admin/finca/{id_finca}/patologia', [App\Http\Controllers\HomeController::class, 'crear_patologia'])->name('patologia.crear');
 
-Route::get('/home/sisga-admin/raza/{id}', [App\Http\Controllers\HomeController::class, 'editar_raza'])->name('raza.editar'); 
+Route::get('/home/sisga-admin/finca/{id_finca}/patologia/{id}', [App\Http\Controllers\HomeController::class, 'editar_patologia'])->name('patologia.editar'); 
 
-Route::put('/home/sisga-admin/raza/{id}', [App\Http\Controllers\HomeController::class, 'update_raza'])->name('raza.update');
+Route::put('/home/sisga-admin/finca/{id_finca}/patologia/{id}', [App\Http\Controllers\HomeController::class, 'update_patologia'])->name('patologia.update');
 
-Route::delete('/home/sisga-admin/raza/eliminar/{id}', [App\Http\Controllers\HomeController::class, 'eliminar_raza'])->name('raza.eliminar'); 
-
-/*-->> End Ruta raza*/
+Route::delete('/home/sisga-admin/finca/{id_finca}/patologia/{id}', [App\Http\Controllers\HomeController::class, 'eliminar_patologia'])->name('patologia.eliminar'); 
 
 
 
 
 /* ----> BeginRoute: Estas son las rutas 
-definidas para el Módulo de ganadería <-----*/
+definidas para el Módulo de ganadería <-----
+
+Route::get('/home/ganaderia/ficha-de-ganado', [App\Http\Controllers\RutasController::class, 'ganaderia'])->name('ficha'); 
+*/
 
 /*Ruta para la vista ficha de ganado*/
-Route::get('/home/ganaderia/ficha-de-ganado', [App\Http\Controllers\RutasController::class, 'ganaderia'])->name('ficha'); 
+Route::get('/home/sisga-admin/finca/{id_finca}/ganaderia/ficha-de-ganado', [App\Http\Controllers\RutasController::class, 'ganaderia'])->name('ficha'); 
 
 	//--> Ruta para filtrar el select Tipologia por sexo en forma dinamica.
 	Route::get('/home/ganaderia/ficha-de-ganado/sexo/{sexo}', [App\Http\Controllers\RutasController::class, 'filterSexo'])->name('filterSexo');
 
-Route::post('/home/ganaderia/ficha-de-ganado', [App\Http\Controllers\RutasController::class, 'crear_fichaganado'])->name('fichaganado.crear');
+Route::post('/home/sisga-admin/finca/{id_finca}/ganaderia/ficha-de-ganado', [App\Http\Controllers\RutasController::class, 'crear_fichaganado'])->name('fichaganado.crear');
 
-Route::get('/home/ganaderia/ficha-de-ganado/{id}', [App\Http\Controllers\RutasController::class, 'editar_fichaganado'])->name('fichaganado.editar');
+Route::get('/home/sisga-admin/finca/{id_finca}/ganaderia/ficha-de-ganado/{id}', [App\Http\Controllers\RutasController::class, 'editar_fichaganado'])->name('fichaganado.editar');
 
-Route::put('/home/ganaderia/ficha-de-ganado/{id}', [App\Http\Controllers\RutasController::class, 'update_fichaganado'])->name('fichaganado.update');  
+Route::put('/home/sisga-admin/finca/{id_finca}/ganaderia/ficha-de-ganado/{id}', [App\Http\Controllers\RutasController::class, 'update_fichaganado'])->name('fichaganado.update');  
 	
 
 
@@ -224,15 +262,19 @@ Route::get('/home/ganaderia/transferencia', [App\Http\Controllers\RutasControlle
 
 
 /*Ruta para la vista de pajuela*/
+
+Route::get('/home/sisga-admin/finca/{id_finca}/ganaderia/pajuela', [App\Http\Controllers\RutasController::class, 'pajuela'])->name('pajuela'); 
+
+/*
 Route::get('/home/ganaderia/pajuela', [App\Http\Controllers\RutasController::class, 'pajuela'])->name('pajuela'); 
+*/
+Route::post('/home/sisga-admin/finca/{id_finca}/ganaderia/pajuela', [App\Http\Controllers\RutasController::class, 'crear_pajuela'])->name('pajuela.crear'); 
 
-Route::post('/home/ganaderia/pajuela', [App\Http\Controllers\RutasController::class, 'crear_pajuela'])->name('pajuela.crear'); 
+Route::get('/home/sisga-admin/finca/{id_finca}/ganaderia/pajuela/{id}', [App\Http\Controllers\RutasController::class, 'editar_pajuela'])->name('pajuela.editar'); 
 
-Route::get('/home/ganaderia/pajuela/{id}', [App\Http\Controllers\RutasController::class, 'editar_pajuela'])->name('pajuela.editar'); 
+Route::put('/home/sisga-admin/finca/{id_finca}/ganaderia/pajuela/{id}', [App\Http\Controllers\RutasController::class, 'update_pajuela'])->name('pajuela.update');
 
-Route::put('/home/ganaderia/pajuela/{id}', [App\Http\Controllers\RutasController::class, 'update_pajuela'])->name('pajuela.update');
-
-Route::delete('/home/ganaderia/pajuela/eliminar/{id}', [App\Http\Controllers\RutasController::class, 'eliminar_pajuela'])->name('pajuela.eliminar'); 
+Route::delete('/home/sisga-admin/finca/{id_finca}/ganaderia/pajuela/eliminar/{id}', [App\Http\Controllers\RutasController::class, 'eliminar_pajuela'])->name('pajuela.eliminar'); 
 
 //->END PAJUELA
